+++ sampler.h
@@ -1,7 +1,6 @@
 #ifndef SAMPLER_H
 #define SAMPLER_H
 
-#include <stddef.h>
 #include "surface.h"
 
 class Sampler
@@ -10,14 +9,14 @@ public:
     Sampler();
     virtual ~Sampler();
 
-    void setSize(size_t width,
-                 size_t height);
+    void setSize(unsigned int width,
+                 unsigned int height);
 
     virtual Surface *createSurface() const = 0;
 
 protected:
-    size_t m_width;
-    size_t m_height;
+    unsigned int m_width;
+    unsigned int m_height;
 };
 
 #endif // SAMPLER_H
