+++ simplejuliasampler.cpp
@@ -15,9 +15,9 @@ Surface *SimpleJuliaSampler::createSurface() const
     surface->setMaxValue(m_steps + 1);
 
 #pragma omp parallel for
-    for (size_t y = 0; y < m_height; y++)
+    for (unsigned int y = 0; y < m_height; y++)
     {
-        for (size_t x = 0; x < m_width; x++)
+        for (unsigned int x = 0; x < m_width; x++)
         {
             Real xPos = m_bottomLeft.real() + xStep * (x + 0.5);
             Real yPos = m_topRight.imag() - yStep * (y + 0.5);
