+++ surface.cpp
@@ -1,7 +1,7 @@
 #include "surface.h"
 
-Surface::Surface(size_t width,
-                 size_t height) :
+Surface::Surface(unsigned int width,
+                 unsigned int height) :
     m_width(width),
     m_height(height),
     m_maxValue(0),
@@ -9,12 +9,12 @@ Surface::Surface(size_t width,
 {
 }
 
-size_t Surface::getWidth() const
+unsigned int Surface::getWidth() const
 {
     return m_width;
 }
 
-size_t Surface::getHeight() const
+unsigned int Surface::getHeight() const
 {
     return m_height;
 }
@@ -29,8 +29,8 @@ void Surface::setMaxValue(Level value)
     m_maxValue = value;
 }
 
-Level Surface::getPoint(size_t x,
-                        size_t y) const
+Level Surface::getPoint(unsigned int x,
+                        unsigned int y) const
 {
     if ((x >= m_width) || (y >= m_height))
     {
@@ -40,8 +40,8 @@ Level Surface::getPoint(size_t x,
     return m_points[y * m_width + x];
 }
 
-bool Surface::setPoint(size_t x,
-                       size_t y,
+bool Surface::setPoint(unsigned int x,
+                       unsigned int y,
                        Level value)
 {
     if ((x >= m_width) || (y >= m_height))
